[BUGIFX] Length of published trajectory reduced to 50 to safe cpu runtime.
As reported by many people, simulator slows down when the camera is switched on and the car can not be controlled anymore. This issue has been identified and investigated by many people and one fix seems to be reducing the length of the trajectory. The fix works for me (my config: super old phenom x6, two cores/4gb ram for win10 where simulator is running and four cores for vm machine with 4 gb ram).
Let me know if the configuration works for you as well!