[PLAYGROUND] Initial commit for object detection. Just playing around, maybe it might be useful later on.
NOTE: There is a heavy bug with unknown reason. A node dies at startup, still everything works?!?.

Coco network is used to identify objects. Works for traffic lights (but only very sparse detection). Needs a lot of runtime, is able to be processed with 2 Hz currently. Copy-paste from object detection lab with some customization. For testing, a new topic was made: "/image_detection", which I subscribed with the rqt_image_viewer to see the results.