[ENHANCEMENT, BUGFIX] Implementation of the improvement of the planner: It does not plan the velocity as well when a red traffic light appears. Bugfix in twist_controller
The planner now takes the latest 80 trajectories (~52 meters) in front of the car to plan the velocity. Once a red traffic light's stopline is reported to the trajectory planner, it creates a linear decrease of the velocity from the point where the car is until the stopline (1 m ahead), where a velocity of 0 m/s is planned. Note that the planner somehow might not be able to stop if the traffic light switches to red when the car is less then 1 m ahead of the TL ("yellow light feature required"). The twist controller is buggy: Previously it was not able to brake (wrong parametrization). Braking works, but the controller uses brakes at constant speed to keep speed (to be optimized)