[ENHANCEMENT] Some minor updates for making the car stop properly (with traffic light status signal).
The red light status filter has been removed and needs to be re-implemented once the image based TL detection is used. As it turns out, the filter is quite slow, since the images are also provided only once per 500 ms or so.
TODO: The planner makes an error log (as debug messages) of the trajectories speed.